Crate juniper_warp

juniper_warp

This repository contains the warp web server integration for Juniper, a GraphQL implementation for Rust.

Documentation

For documentation, including guides and examples, check out Juniper.

A basic usage example can also be found in the Api documentation.

Examples

Check examples/warp_server for example code of a working warp server with GraphQL handlers.

Links

Juniper

Api Reference

warp

License

This project is under the BSD-2 license.

Check the LICENSE file for details.

Functions

graphiql_filter
Create a filter that replies with an HTML page containing GraphiQL. This does not handle routing, so you can mount it on any endpoint.

make_graphql_filter
Make a filter for graphql endpoint.

make_graphql_filter_async
FIXME: docs

playground_filter
Create a filter that replies with an HTML page containing GraphQL Playground. This does not handle routing, so you can mount it on any endpoint.